Move attr struct to own header
